MLE on the DFE for FGM

Maximum likelihood estimation (mle) for the parameters of the DFE density for different form of the FGM.

mle_dfgm(s, model = "Martin", start, method = "NM", ...)

Arguments

s
Vector of real numbers. Quantiles of the density. The density is 0 if \(s \ge so\).

model
String. "Martin" performs the mle on the density of the DFE from Martin & Lenormand (2015) (see dfgm_martin). "Tenaillon" performs the mle on the density of the DFE from Tenaillon (2014) (see dfgm_tenaillon). Default "Martin".

start
Numeric vector (named). Used as starting values in maxLik{maxLik}. It must contains the parameters for the chosen model:

"Martin" c(n = #, lambda = #, so = #)

"Tenaillon" c(n = #, lambda = #, so = #, alpha = #, Q = #)

method
String. Maximisation method. See maxLik{maxLik}. Default "NM".

...
See maxLik{maxLik} all the possible parameters

Source

Tenaillon, O. (2014). The utility of Fisher's geometric model in evolutionary genetics. Annual review of ecology, evolution, and systematics, 45, 179-201.

Martin, G., & Lenormand, T. (2015). The fitness effect of mutations across environments: Fisher's geometrical model with multiple optima. Evolution, 69(6), 1433-1447.

Value

Returns the output of maxLik for the chosen model.

See also

Examples

#### model : "Martin" ####
# Parameters
nsample <- 1000; n <- 4; lambda <- 0.5; so <- 1; alpha <- 1/2; Q <- 2;
# Simulated data
s <- rfgm(nsample, n, lambda, so, alpha, Q)
# Noise on initial parameters
initial_parameter <- c(n = ceiling(abs(rnorm(1, n))),
                       lambda = abs(rnorm(1, lambda, sd = 0.1)),
                       so = abs(rnorm(1, n)))
# Constraints on parameters (not applioed for all methods see maxLik in maxLik package)
consA <- rbind(c(1, 0, 0),
               c(0, 1, 0),
               c(0, 0, 1))
consB <- c(0, 0, -max(s))
# MLE
res <- mle_dfgm(s, model = "Martin", start = initial_parameter,
                method = "NM", constraints = list(ineqA = consA, ineqB = consB))

#### model : "Tenaillon" ####
# Parameters
nsample <- 1000; n <- 4; lambda <- 0.5; so <- 1; alpha <- 1/2; Q <- 2;
# Simulated data
s <- rfgm(nsample, n, lambda, so, alpha, Q)
# Noise on initial parameters
initial_parameter <- c(n = ceiling(abs(rnorm(1, n))),
                       lambda = abs(rnorm(1, lambda, sd = 0.1)),
                       so = abs(rnorm(1, n)),
                       alpha = alpha,
                       Q = abs(rnorm(1, Q)))
# Constraints on parameters (not applioed for all methods see maxLik in maxLik package)
consA <- rbind(c(1, 0, 0, 0, 0),
               c(0, 1, 0, 0, 0),
               c(0, 0, 1, 0, 0),
               c(0, 0, 0, 1, 0),
               c(0, 0, 0, 0, 1))
consB <- c(0, 0, -max(s), 0, 0)
# MLE
res <- mle_dfgm(s, model = "Tenaillon", start = initial_parameter,
                method = "NM", constraints = list(ineqA = consA, ineqB = consB))